Place the same pom with HubSpot version; Move classes under the same package
