+++ src/main/java/de/moritzf/nlp/jtopia/entities/HubSpotStyle.java
@@ -1,0 +1,4 @@
+package de.moritzf.nlp.jtopia.entities;
+
+public class HubSpotStyle {
+}
